+++ cdlclient/tests/connect_dialog.py
@@ -10,17 +10,21 @@ DataLab Remote client connection dialog example
 # guitest: show
 
 from guidata.qthelpers import qt_app_context
+from qtpy import QtWidgets as QW
 
 from cdlclient import SimpleRemoteProxy
-from cdlclient.widgets import DataLabConnectionDialog
+from cdlclient.widgets import ConnectionDialog
 
 
 def test_dialog():
     """Test connection dialog"""
     proxy = SimpleRemoteProxy()
     with qt_app_context():
-        dlg = DataLabConnectionDialog(proxy.connect)
-        dlg.exec()
+        dlg = ConnectionDialog(proxy.connect)
+        if dlg.exec():
+            QW.QMessageBox.information(None, "Connection", "Successfully connected")
+        else:
+            QW.QMessageBox.critical(None, "Connection", "Connection failed")
 
 
 if __name__ == "__main__":